chore(task): adds task info:booster:pools-tvl
adds task info:booster:pools-tvl to calculate migrated tvls on pools with old stash version
